# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Agent {
  id: number;
  slug: string;
  name: string;
  shortDescription: string;
  order: number;
}

export interface SampleExample {
  title: string;
  input: string;
  output: string;
}

export interface AgentDetail {
  id: number;
  slug: string;
  name: string;
  shortDescription: string;
  description: string;
  howItWorks: string;
  requirements: string;
  expectedOutput: string;
  sampleExamples: SampleExample[];
  order: number;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}