# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: agents
    description: AI Agent operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      description: Returns all available AI agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"

  /agents/{slug}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by slug
      description: Returns a single agent's full details
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      description: Authenticate as admin with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      description: Destroy the current session
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current admin
      description: Returns the currently authenticated admin or 401
      responses:
        "200":
          description: Current admin user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Agent:
      type: object
      required: [id, slug, name, shortDescription, order]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        order:
          type: integer

    SampleExample:
      type: object
      required: [title, input, output]
      properties:
        title:
          type: string
        input:
          type: string
        output:
          type: string

    AgentDetail:
      type: object
      required: [id, slug, name, shortDescription, description, howItWorks, requirements, expectedOutput, sampleExamples, order]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        howItWorks:
          type: string
        requirements:
          type: string
        expectedOutput:
          type: string
        sampleExamples:
          type: array
          items:
            $ref: "#/components/schemas/SampleExample"
        order:
          type: integer

    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AdminUser:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string