# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Agent {
  id: number;
  slug: string;
  name: string;
  shortDescription: string;
  order: number;
}