# lib/api-zod/src/generated/types/agentDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SampleExample } from './sampleExample';

export interface AgentDetail {
  id: number;
  slug: string;
  name: string;
  shortDescription: string;
  description: string;
  howItWorks: string;
  requirements: string;
  expectedOutput: string;
  sampleExamples: SampleExample[];
  order: number;
}